requirements.txt
python-dotenv===1.0.0
discord.py===2.2.2
requests===2.28.2
tabulate===0.9.0